fix: 실험 폴더명 타임스탬프 누락 및 LLM BFloat16 오류 수정
1. 실험 폴더명에 날짜_시간_ 접두사 항상 추가
   - 사용자 지정 experiment_name 사용 시에도 타임스탬프 접두사 적용
   - 폴더명 형식: YYYYMMDD_HHMMSS_{experiment_name}

2. Logger 진행률 표시줄 중복 방지 강화
   - ANSI 이스케이프 시퀀스 제거 후 중복 감지
   - 진행률 지표 확장 ('|', '[00:' 추가)
   - 지표와 숫자 동시 존재 시에만 진행률로 간주

3. LLM Quantization BFloat16 기본값 수정
   - bnb_4bit_compute_dtype 기본값 변경: 'bfloat16' → 'float16'
   - PyTorch AMP gradient scaling BFloat16 미지원 오류 해결
   - llama-3.2-korean-3b 등 LLM 모델 학습 안정성 개선

+++ src/models/llm_loader.py
@@ -45,7 +45,7 @@ def load_causal_lm(config, logger=None):
             load_in_4bit=config.model.quantization.get('load_in_4bit', True),
             bnb_4bit_compute_dtype=getattr(
                 torch,
-                config.model.quantization.get('bnb_4bit_compute_dtype', 'bfloat16')
+                config.model.quantization.get('bnb_4bit_compute_dtype', 'float16')
             ),
             bnb_4bit_quant_type=config.model.quantization.get('bnb_4bit_quant_type', 'nf4'),
             bnb_4bit_use_double_quant=config.model.quantization.get('bnb_4bit_use_double_quant', True)
